% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{check_service}
\alias{check_service}
\title{Check whether a web service is unavailable, and stop function if necessary.}
\usage{
check_service(x)
}
\arguments{
\item{x}{The path to the web service.}
}
\value{
Error if service unavailable.
}
\description{
Check whether a web service is unavailable, and stop function if necessary.
}
\keyword{internal}
