% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHCN_FUNCTIONS.R
\name{download_ghcn_daily_station}
\alias{download_ghcn_daily_station}
\title{Download the daily data for a GHCN weather station.}
\usage{
download_ghcn_daily_station(ID, raw.dir, force.redo = FALSE)
}
\arguments{
\item{ID}{A character string giving the station ID.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.}

\item{force.redo}{If this weather station has been downloaded before, should it be updated? Defaults to FALSE.}
}
\value{
A character string representing the full local path of the GHCN station data.
}
\description{
Download the daily data for a GHCN weather station.
}
\keyword{internal}
