% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_FUNCTIONS.R
\name{download_nhd_subregion}
\alias{download_nhd_subregion}
\title{Download a zipped NHD HUC4 subregion.}
\usage{
download_nhd_subregion(area, res, raw.dir)
}
\arguments{
\item{area}{A 4-character string indicating the HUC4 NHD subregion to download.}

\item{res}{A character string defining the resolution of the NHD to download.
Either "medium" (the default), or "high".}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing.}
}
\value{
A character string representing the full local path of the downloaded zip file.
}
\description{
HUC4 subregion are specified by a unique character string, best obtained using the \code{\link{get_huc4}} function.
\code{download_nhd_subregion} returns the path to the downloaded zip file.
}

