% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{download_ssurgo_inventory}
\alias{download_ssurgo_inventory}
\title{Download a zipped directory containing a shapefile of the SSURGO study areas.}
\usage{
download_ssurgo_inventory(raw.dir)
}
\arguments{
\item{raw.dir}{A character string indicating where raw downloaded files should be put.}
}
\value{
A character string representing the full local path of the SSURGO study areas zipped directory.
}
\description{
Download a zipped directory containing a shapefile of the SSURGO study areas.
}
\keyword{internal}
