% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{download_ssurgo_study_area}
\alias{download_ssurgo_study_area}
\title{Download a zipped directory containing the spatial and tabular data for a SSURGO study area.}
\usage{
download_ssurgo_study_area(area, date, raw.dir)
}
\arguments{
\item{area}{A character string indicating the SSURGO study area to be downloaded.}

\item{date}{A character string indicating the date of the most recent update to the SSURGO 
area for these data. This information may be gleaned from the SSURGO Inventory (\code{\link{get_ssurgo_inventory}}).}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.}
}
\value{
A character string representing the full local path of the SSURGO study areas zipped directory.
}
\description{
\code{download_ssurgo_study_area} first tries to download data including a state-specific Access
template, then the general US template.
}
\keyword{internal}

