% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NHD_FUNCTIONS.R
\name{downloadHUC4}
\alias{downloadHUC4}
\title{Download a zipped directory containing a shapefile of the HUC4 subregions of the NHD.}
\usage{
downloadHUC4(raw.dir)
}
\arguments{
\item{raw.dir}{A character string indicating where raw downloaded files should be put.}
}
\value{
A character string representing the full local path of the HUC4 zipped directory.
}
\description{
Download a zipped directory containing a shapefile of the HUC4 subregions of the NHD.
}

