% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NHD_FUNCTIONS.R
\name{downloadNHDSubregion}
\alias{downloadNHDSubregion}
\title{Download a zipped NHD HUC4 subregion.}
\usage{
downloadNHDSubregion(area, raw.dir)
}
\arguments{
\item{area}{A 4-character string indicating the HUC4 NHD subregion to download.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing.}
}
\value{
A character string representing the full local path of the downloaded zip file.
}
\description{
HUC4 subregion are specified by a unique character string, best obtained using the \code{\link{getHUC4}} function.
\code{downloadNHDSubregion} returns the path to the downloaded zip file.
}

