% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{extract_ssurgo_data}
\alias{extract_ssurgo_data}
\title{Extract data from a SSURGO databse pertaining to a set of mapunits.}
\usage{
extract_ssurgo_data(tables, mapunits)
}
\arguments{
\item{tables}{A list of SSURGO tabular data.}

\item{mapunits}{A character vector of mapunits (likely dropped from SSURGO spatial data)
defining which mapunits to retain.}
}
\value{
A list of extracted SSURGO tabular data.
}
\description{
\code{extract_ssurgo_data} creates a directed graph of the joins in a SSURGO tabular dataset,
and then iterates through the tables, only retaining data pertinant to a set of mapunits.
}
\keyword{internal}
