% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{extractSSURGOData}
\alias{extractSSURGOData}
\title{Extract data from a SSURGO databse pertaining to a set of mapunits.}
\usage{
extractSSURGOData(tables, mapunits)
}
\arguments{
\item{tables}{A list of SSURGO tabular data.}

\item{mapunits}{A \code{SpatialPolygonsDataFrame} of mapunits (likely dropped from SSURGO spatial data)
defining which mapunits to retain.}
}
\value{
A list of extracted SSURGO tabular data.
}
\description{
\code{extractSSURGOData} creates a directed graph of the joins in a SSURGO tabular dataset,
and then iterates through the tables, only retaining data pertinant to a set of mapunits.
}

