\name{FedData-package}
\alias{FedData-package}
\alias{FedData}
\docType{package}
\title{
Scripts to automate downloading geospatial data available from the several federated data sources
}
\description{
Scripts to automate downloading geospatial data available from the several federated data sources.
}
\details{
\tabular{ll}{
Package: \tab FedData\cr
Type: \tab FedData\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-09\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
R. Kyle Bocinsky <bocinsky@gmail.com>
}
\keyword{ package }