% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_FUNCTIONS.R
\name{get_huc4}
\alias{get_huc4}
\title{Download and crop a shapefile of the HUC4 
regions of the National Hydrography Dataset.}
\usage{
get_huc4(template = NULL, raw.dir)
}
\arguments{
\item{template}{A Raster* or Spatial* object to serve 
as a template for cropping.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing.}
}
\value{
A \code{SpatialPolygonsDataFrame} of the HUC4 regions within
the specified \code{template}.
}
\description{
\code{get_huc4} returns a \code{SpatialPolygonsDataFrame} of the HUC4 regions within
the specified \code{template}. If template is not provided, returns the entire HUC4 dataset.
}
\keyword{internal}
