% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCD_FUNCTIONS.R
\name{get_nlcd}
\alias{get_nlcd}
\title{Download and crop the National Land Cover Database.}
\usage{
get_nlcd(
  template,
  label,
  year = 2016,
  dataset = "Land_Cover",
  landmass = "L48",
  extraction.dir = paste0(tempdir(), "/FedData/"),
  force.redo = F
)
}
\arguments{
\item{template}{A Raster* or Spatial* object to serve
as a template for cropping.}

\item{label}{A character string naming the study area.}

\item{year}{An integer representing the year of desired NLCD product.
Acceptable values are 2016 (default), 2011, 2008 (landcover only), 2006, 2004, and 2001.}

\item{dataset}{A character string representing type of the NLCD product.
Acceptable values are 'Impervious', 'Land_Cover', 'Tree_Canopy' (2011 and 2016 only),}

\item{landmass}{A character string representing the landmass to be extracted
Acceptable values are 'L48' (lower 48 US states, the default), 'AK' (Alaska), 'HI' (Hawaii), and 'PR' (Puerto Rico).}

\item{extraction.dir}{A character string indicating where the extracted and cropped NLCD data should be put.
The directory will be created if missing.}

\item{force.redo}{If an extraction for this template and label already exists, should a new one be created?}
}
\value{
A \code{RasterLayer} cropped to the bounding box of the template.
}
\description{
\code{get_nlcd} returns a \code{RasterLayer} of NLCD data cropped to a given
template study area.
}
\examples{
\dontrun{
# Extract data for the Mesa Verde National Park:

# Get the NLCD (USA ONLY)
# Returns a raster
NLCD <- get_nlcd(paleocar::mvnp \%>\%
  sf::st_as_sf(),
label = "MVNP",
year = 2011,
landmass = "L48"
)

# Plot with raster::plot
plot(NLCD)
}
}
