% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{get_ssurgo_inventory}
\alias{get_ssurgo_inventory}
\title{Download and crop a shapefile of the SSURGO study areas.}
\usage{
get_ssurgo_inventory(template = NULL, raw.dir)
}
\arguments{
\item{template}{An \code{\link[sf:sf]{Simple Feature}}
or \code{\link[terra:SpatRaster-class]{SpatRaster}} object to serve as a template for cropping.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing.}
}
\value{
A \code{SpatialPolygonsDataFrame} of the SSURGO study areas within
the specified \code{template}.
}
\description{
\code{get_ssurgo_inventory} returns a \code{SpatialPolygonsDataFrame} of the SSURGO study areas within
the specified \code{template}. If template is not provided, returns the entire SSURGO inventory of study areas.
}
\keyword{internal}
