% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NED_FUNCTIONS.R
\name{getNEDTile}
\alias{getNEDTile}
\title{Download and crop tile from the 1 (~30 meter) or 1/3 (~10 meter) arc-second National Elevation Dataset.}
\usage{
getNEDTile(template = NULL, res, tileNorthing, tileWesting, raw.dir)
}
\arguments{
\item{template}{A Raster* or Spatial* object to serve
as a template for cropping.}

\item{res}{A character string representing the desired resolution of the NED. "1"
indicates the 1 arc-second NED, while "13" indicates the 1/3 arc-second dataset. Defaults to NULL.}

\item{tileNorthing}{An integer representing the northing (latitude, in degrees north of the equator) of the northwest corner of the tile to
be downloaded.}

\item{tileWesting}{An integer representing the westing (longitude, in degrees west of the prime meridian) of the northwest corner of the tile to
be downloaded.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing. Defaults to "./RAW/NED/".}
}
\value{
A \code{RasterLayer} cropped within the specified \code{template}.
}
\description{
\code{getNEDTile} returns a \code{RasterLayer} cropped within the specified \code{template}.
If template is not provided, returns the entire NED tile.
}

