% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{getSSURGOInventory}
\alias{getSSURGOInventory}
\title{Download and crop a shapefile of the SSURGO study areas.}
\usage{
getSSURGOInventory(template = NULL, raw.dir)
}
\arguments{
\item{template}{A Raster* or Spatial* object to serve
as a template for cropping.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing.}
}
\value{
A \code{SpatialPolygonsDataFrame} of the SSURGO study areas within
the specified \code{template}.
}
\description{
\code{getSSURGOInventory} returns a \code{SpatialPolygonsDataFrame} of the SSURGO study areas within
the specified \code{template}. If template is not provided, returns the entire SSURGO inventory of study areas.
}

