% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glac}
\alias{glac}
\title{The boundary of Glacier National Park}
\format{
Simple feature collection with 1 feature and 5 fields.
}
\usage{
glac
}
\description{
A dataset containing the spatial polygon defining the boundary
of Glacier National Park in Montana.
}
\keyword{datasets}
