% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meve}
\alias{meve}
\title{The boundary of Mesa Verde National Park}
\format{
Simple feature collection with 1 feature and a geometry field.
}
\usage{
meve
}
\description{
A dataset containing the spatial polygon defining the boundary
of Mesa Verde National Park in Montana.
}
\keyword{datasets}
