% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\docType{data}
\name{mvnp}
\alias{mvnp}
\title{The boundary of Mesa Verde National Park}
\format{
Simple feature collection with 1 feature and 10 fields.
}
\usage{
mvnp
}
\description{
A dataset containing the spatial polygon defining the boundary 
of Mesa Verde National Park in southwestern Colorado.
}
\keyword{datasets}
