% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCD_FUNCTIONS.R
\docType{data}
\name{nlcd_impervious_pam}
\alias{nlcd_impervious_pam}
\title{The NLCD impervious PAM attributes.}
\format{An object of class \code{character} of length 2345.}
\usage{
nlcd_impervious_pam
}
\description{
A dataset containing the PAM attributes.
}
\keyword{datasets}
