% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCD_FUNCTIONS.R
\docType{data}
\name{nlcd_tiles}
\alias{nlcd_tiles}
\title{The NLCD tiles SpatialPolygonsDataFrame.}
\format{A SpatialPolygonsDataFrame with 203 features and 1 variable:
\describe{
  \item{Name}{the name of the tile}
}}
\usage{
nlcd_tiles
}
\description{
A dataset containing the NLCD tiles.
}
\keyword{internal}
