% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{pkgTest}
\alias{pkgTest}
\title{Install and load a package.}
\usage{
pkgTest(x)
}
\arguments{
\item{x}{A character string representing the name of a package.}
}
\description{
This is a convenience function that checks whether a package is installed, and if not, installs it.
}

