% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHD_FUNCTIONS.R
\name{plot_nhd}
\alias{plot_nhd}
\title{A basic plotting function for NHD data.}
\usage{
plot_nhd(x, template = NULL)
}
\arguments{
\item{x}{The result of \link{get_nhd}.}

\item{template}{An \code{\link[sf:sf]{Simple Feature}}
or \code{\link[terra:SpatRaster-class]{SpatRaster}} object to serve as a template for cropping.}
}
\value{
A \code{ggplot2} panel of plots
}
\description{
This is more of an example than anything
}
\examples{
\donttest{
# Get the NHD (USA ONLY)
NHD <- get_nhd(
  template = FedData::meve,
  label = "meve"
)
NHD
NHD \%>\%
  plot_nhd(template = FedData::meve)
}
}
