% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULLs}
\usage{
replace_null(x)
}
\arguments{
\item{x}{A list}
}
\description{
Replace all the empty values in a list
}
\examples{
list(a = NULL, b = 1, c = list(foo = NULL, bar = NULL)) \%>\% replace_null()
}
