% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{SDA_query}
\alias{SDA_query}
\title{Submit a Soil Data Access (SDA) Query}
\usage{
SDA_query(q)
}
\arguments{
\item{q}{A character string representing a SQL query to the SDA service}
}
\value{
A tibble returned from the SDA service
}
\description{
\code{SDA_query} submit an SQL or SQL Data Shaping query to retrieve data from the Soil Data Mart.
Please see https://sdmdataaccess.sc.egov.usda.gov/Query.aspx for guidelines
}
\keyword{internal}
