% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{sequential.duplicated}
\alias{sequential.duplicated}
\title{Get a logical vector of which elements in a vector are sequentially duplicated.}
\usage{
sequential.duplicated(x, rows = F)
}
\arguments{
\item{x}{An vector of any type, or, if \code{rows}, a matrix.}

\item{rows}{Is x a matrix?}
}
\value{
A logical vector of the same length as x.
}
\description{
Get a logical vector of which elements in a vector are sequentially duplicated.
}

