% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_FUNCTIONS.R
\name{soils_query}
\alias{soils_query}
\title{Submit a Soil Data Access (SDA) Query}
\usage{
soils_query(q)
}
\arguments{
\item{q}{A character string representing a SQL query to the SDA service}
}
\value{
A tibble returned from the SDA service
}
\description{
\code{soils_query} submit an SQL query to retrieve data from the Soil Data Mart.
Please see https://sdmdataaccess.sc.egov.usda.gov/Query.aspx for guidelines
}
\keyword{internal}
