% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{SPDFfromPolygon}
\alias{SPDFfromPolygon}
\title{Turn an SpatialPolygons object into a SpatialPolygonsDataFrame.}
\usage{
SPDFfromPolygon(x)
}
\arguments{
\item{x}{An SpatialPolygons object.}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
Turn an SpatialPolygons object into a SpatialPolygonsDataFrame.
}

