% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{split_bbox}
\alias{split_bbox}
\title{Splits a bbox into a list of bboxes less than a certain size}
\usage{
split_bbox(bbox, x, y = x)
}
\arguments{
\item{x}{The maximum x size of the resulting bounding boxes}

\item{y}{The maximum y size of the resulting bounding boxes; defaults to x}
}
\value{
A list of bbox objects
}
\description{
Splits a bbox into a list of bboxes less than a certain size
}
\keyword{internal}
