% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{substr_right}
\alias{substr_right}
\title{Get the rightmost 'n' characters of a character string.}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{A character string.}

\item{n}{The number of characters to retrieve.}
}
\value{
A character string.
}
\description{
Get the rightmost 'n' characters of a character string.
}
\keyword{internal}
