% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{substrRight}
\alias{substrRight}
\title{Get the rightmost "n" characters of a character string.}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{A character string.}

\item{n}{The number of characters to retrieve.}
}
\value{
A character string.
}
\description{
Get the rightmost "n" characters of a character string.
}

