% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NED_FUNCTIONS.R
\name{unzipNED}
\alias{unzipNED}
\title{Unzip a zipped tile from the 1 (~30 meter) or 1/3 (~10 meter) arc-second National Elevation Dataset.}
\usage{
unzipNED(file)
}
\arguments{
\item{file}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing. Defaults to "./RAW/NED/".}
}
\value{
A character string representing the full local path of a downloaded zip file to be decompressed.
}
\description{
\code{unzipNED} decompresses a downloaded NED tile and returns a \code{\link[raster]{RasterLayer}}
of the extracted tile. The raster layer is held completely in memory.
}

