% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{url_base}
\alias{url_base}
\title{Strip query parameters from a URL}
\usage{
url_base(x)
}
\arguments{
\item{x}{The URL to be modified}
}
\value{
The URL without parameters
}
\description{
Strip query parameters from a URL
}
\keyword{internal}
