% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{wgetDownload}
\alias{wgetDownload}
\title{Use the \code{wget} command line tool to download a file.}
\usage{
wgetDownload(url, destdir = getwd(), timestamping = T, nc = F)
}
\arguments{
\item{url}{The location of a file.}

\item{destdir}{Where the file should be downloaded to.}

\item{timestamping}{Should only newer files be downloaded?}

\item{nc}{Should files of the same type not be clobbered?}
}
\value{
A logical vector of the same length as x.
}
\description{
If both \code{timestamping} and \code{nc} are TRUE, timestamping behavior trumps nc.
}

