% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{polygonFromExtent}
\alias{polygonFromExtent}
\title{Turn an extent object into a polygon}
\usage{
polygonFromExtent(x, proj4string = NULL)
}
\arguments{
\item{x}{An \code{\link{extent}} object, or an object from which an extent object can be retrieved.}

\item{proj4string}{A PROJ.4 formatted string defining the required projection. If NULL,
the function will attempt to get the projection from x using \code{\link{projection}}}
}
\value{
A SpatialPolygons object.
}
\description{
Turn an extent object into a polygon
}

