% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILITY_FUNCTIONS.R
\name{spdf_from_polygon}
\alias{spdf_from_polygon}
\title{Turn a SpatialPolygons object into a SpatialPolygonsDataFrame.}
\usage{
spdf_from_polygon(x)
}
\arguments{
\item{x}{An SpatialPolygons object.}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
Turn a SpatialPolygons object into a SpatialPolygonsDataFrame.
}
\keyword{internal}

